# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgriLog v2 API",
    "description": "API cho ứng dụng nông hộ AgriLog v2 (mobile) và trang quản trị (web-admin).\n\n**Xác thực:** đăng nhập ở `POST /auth/login`, rồi gửi `Authorization: Bearer <token>` cho mọi endpoint khác.\n\n**Đồng bộ:** `GET /sync` kéo thay đổi, `POST /sync` đẩy thay đổi. Xung đột giải quyết theo *last write wins* dựa trên `updated_at`, và mọi bản ghi bị giữ lại đều trả về trong `conflicts` kèm bản của máy chủ.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks-history/{task_id}/notes": {
      "get": {
        "tags": [
          "care-protocols"
        ],
        "summary": "List Notes",
        "operationId": "list_notes_tasks_history__task_id__notes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskNoteOut"
                  },
                  "title": "Response List Notes Tasks History  Task Id  Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "care-protocols"
        ],
        "summary": "Create Note",
        "description": "A note on how the task was done. Photos/videos go up first through\nPOST /media (or afterwards through .../upload-media) and are named here.",
        "operationId": "create_note_tasks_history__task_id__notes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskNoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskNoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks-history/{task_id}/notes/{note_id}/upload-media": {
      "post": {
        "tags": [
          "care-protocols"
        ],
        "summary": "Upload Note Media",
        "description": "Uploads one photo/video and attaches it to the note in one call.",
        "operationId": "upload_note_media_tasks_history__task_id__notes__note_id__upload_media_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_note_media_tasks_history__task_id__notes__note_id__upload_media_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskNoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/task-notes/{note_id}": {
      "delete": {
        "tags": [
          "care-protocols"
        ],
        "summary": "Delete Note",
        "operationId": "delete_note_task_notes__note_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks-history/{task_id}/labor-costs": {
      "get": {
        "tags": [
          "care-protocols"
        ],
        "summary": "List Labor Costs",
        "operationId": "list_labor_costs_tasks_history__task_id__labor_costs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaborCostOut"
                  },
                  "title": "Response List Labor Costs Tasks History  Task Id  Labor Costs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "care-protocols"
        ],
        "summary": "Create Labor Cost",
        "description": "Books hired work for the task as an expense of kind \"labor\", so it shows\nup in Thu – Chi, the monthly report and the plot's cost summary.",
        "operationId": "create_labor_cost_tasks_history__task_id__labor_costs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaborCostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborCostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/labor-costs/{cost_id}": {
      "patch": {
        "tags": [
          "care-protocols"
        ],
        "summary": "Update Labor Cost",
        "operationId": "update_labor_cost_labor_costs__cost_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cost_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cost Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaborCostUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborCostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "care-protocols"
        ],
        "summary": "Delete Labor Cost",
        "operationId": "delete_labor_cost_labor_costs__cost_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cost_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cost Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_note_media_tasks_history__task_id__notes__note_id__upload_media_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_note_media_tasks_history__task_id__notes__note_id__upload_media_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LaborCostCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Description"
          },
          "unit": {
            "$ref": "#/components/schemas/LaborUnit",
            "default": "hour"
          },
          "workers": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Workers"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "maximum": 1000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Unit Price"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "description",
          "unit_price"
        ],
        "title": "LaborCostCreate",
        "description": "Hired work for one task.\n\nBy the hour or day: `workers` × `quantity` × `unit_price`. Khoán (a lump\nsum for the whole job): `unit_price` is the sum and the other two are not\nused. The total is always computed here, never taken from the client."
      },
      "LaborCostOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "workers": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workers"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "occurred_at": {
            "type": "integer",
            "title": "Occurred At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "plot_id",
          "description",
          "unit",
          "workers",
          "quantity",
          "unit_price",
          "amount",
          "occurred_at",
          "note",
          "owner_id",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "LaborCostOut"
      },
      "LaborCostUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LaborUnit"
              },
              {
                "type": "null"
              }
            ]
          },
          "workers": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Workers"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "LaborCostUpdate"
      },
      "LaborUnit": {
        "type": "string",
        "enum": [
          "hour",
          "day",
          "lump"
        ],
        "title": "LaborUnit",
        "description": "How hired work is priced: by the hour, by the day (ngày công), or as\none agreed sum for the whole job (khoán)."
      },
...
      "MediaRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          },
          "uploaded": {
            "type": "boolean",
            "title": "Uploaded",
            "default": false
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind"
        ],
        "title": "MediaRef"
      },
...
      "TaskNoteCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "note_text": {
            "type": "string",
            "maxLength": 4000,
            "title": "Note Text",
            "default": ""
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "media_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Media Ids"
          }
        },
        "type": "object",
        "title": "TaskNoteCreate"
      },
      "TaskNoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "note_text": {
            "type": "string",
            "title": "Note Text"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaRef"
            },
            "type": "array",
            "title": "Media"
          },
          "occurred_at": {
            "type": "integer",
            "title": "Occurred At"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "plot_id",
          "note_text",
          "media",
          "occurred_at",
          "owner_id",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "TaskNoteOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}